+++ src/components/MultiHopTrade/components/MultiHopTradeConfirm/hooks/useTradeExecution.tsx
@@ -18,13 +18,10 @@ import {
   getHopByIndex,
   isExecutableTradeQuote,
   SwapperName,
-  swappers,
   TradeExecutionEvent,
 } from '@shapeshiftoss/swapper'
 import { LIFI_TRADE_POLL_INTERVAL_MILLISECONDS } from '@shapeshiftoss/swapper/dist/swappers/LifiSwapper/LifiSwapper'
 import type { CosmosSdkChainId } from '@shapeshiftoss/types'
-import { skipToken, useQuery } from '@tanstack/react-query'
-import { getConfig } from 'config'
 import type { TypedData } from 'eip-712'
 import { useCallback, useEffect, useMemo, useRef } from 'react'
 import { useTranslate } from 'react-polyglot'
@@ -119,178 +116,6 @@ export const useTradeExecution = (
   const supportedBuyAsset = useAppSelector(state =>
     selectAssetById(state, tradeQuote?.steps[hopIndex]?.buyAsset.assetId ?? ''),
   )
-
-  const hop = useMemo(() => getHopByIndex(tradeQuote, hopIndex), [tradeQuote, hopIndex])
-  const swapper = useMemo(() => (swapperName ? swappers[swapperName] : undefined), [swapperName])
-
-  const quoteNetworkFeesCryptoBaseUnitQuery = useQuery({
-    queryKey: ['quoteNetworkFeesCryptoBaseUnit', tradeQuote],
-    queryFn:
-      wallet &&
-      accountMetadata &&
-      tradeQuote &&
-      swapperName &&
-      sellAssetAccountId &&
-      swapper &&
-      hop &&
-      isExecutableTradeQuote(tradeQuote)
-        ? async () => {
-            const execution = new TradeExecution()
-
-            const { accountType, bip44Params } = accountMetadata
-            const accountNumber = bip44Params.accountNumber
-            const stepSellAssetChainId = hop.sellAsset.chainId
-            const stepSellAssetAssetId = hop.sellAsset.assetId
-            const stepBuyAssetAssetId = hop.buyAsset.assetId
-
-            if (swapperName === SwapperName.CowSwap) {
-              // No network fees for CowSwap as this is a message signing
-              return '0'
-            }
-
-            const { chainNamespace: stepSellAssetChainNamespace } =
-              fromChainId(stepSellAssetChainId)
-
-            const receiverAddress =
-              stepBuyAssetAssetId === bchAssetId
-                ? tradeQuote.receiveAddress.replace('bitcoincash:', '')
-                : tradeQuote.receiveAddress
-
-            switch (stepSellAssetChainNamespace) {
-              case CHAIN_NAMESPACE.Evm: {
-                if (!swapper.getEvmTransactionFees) throw Error('missing getEvmTransactionFees')
-                const adapter = assertGetEvmChainAdapter(stepSellAssetChainId)
-                const from = await adapter.getAddress({ accountNumber, wallet })
-                const supportsEIP1559 = supportsETH(wallet) && (await wallet.ethSupportsEIP1559())
-
-                debugger
-                const output = await swapper.getEvmTransactionFees({
-                  chainId: hop.sellAsset.chainId,
-                  tradeQuote,
-                  stepIndex: hopIndex,
-                  slippageTolerancePercentageDecimal,
-                  from,
-                  supportsEIP1559,
-                  permit2Signature: permit2.permit2Signature,
-                  config: getConfig(),
-                  assertGetEvmChainAdapter,
-                })
-                return output
-              }
-              case CHAIN_NAMESPACE.Utxo: {
-                if (accountType === undefined) throw Error('Missing UTXO account type')
-
-                const adapter = assertGetUtxoChainAdapter(stepSellAssetChainId)
-                const { xpub } = await adapter.getPublicKey(wallet, accountNumber, accountType)
-                const _senderAddress = await adapter.getAddress({
-                  accountNumber,
-                  accountType,
-                  wallet,
-                })
-                const senderAddress =
-                  stepSellAssetAssetId === bchAssetId
-                    ? _senderAddress.replace('bitcoincash:', '')
-                    : _senderAddress
-
-                const output = await execution.execUtxoTransaction({
-                  swapperName,
-                  tradeQuote,
-                  stepIndex: hopIndex,
-                  slippageTolerancePercentageDecimal,
-                  xpub,
-                  senderAddress: _senderAddress,
-                  accountType,
-                  signAndBroadcastTransaction: async (txToSign: BTCSignTx) => {
-                    const signedTx = await adapter.signTransaction({
-                      txToSign,
-                      wallet,
-                    })
-
-                    const output = await adapter.broadcastTransaction({
-                      senderAddress,
-                      receiverAddress,
-                      hex: signedTx,
-                    })
-
-                    return output
-                  },
-                })
-                cancelPollingRef.current = output?.cancelPolling
-                return
-              }
-              case CHAIN_NAMESPACE.CosmosSdk: {
-                const adapter = assertGetCosmosSdkChainAdapter(stepSellAssetChainId)
-                const from = await adapter.getAddress({ accountNumber, wallet })
-                const output = await execution.execCosmosSdkTransaction({
-                  swapperName,
-                  tradeQuote,
-                  stepIndex: hopIndex,
-                  slippageTolerancePercentageDecimal,
-                  from,
-                  signAndBroadcastTransaction: async (transactionRequest: StdSignDoc) => {
-                    const txToSign: SignTx<CosmosSdkChainId> = {
-                      addressNList: toAddressNList(adapter.getBip44Params(bip44Params)),
-                      tx: {
-                        fee: {
-                          amount: [...transactionRequest.fee.amount],
-                          gas: transactionRequest.fee.gas,
-                        },
-                        memo: transactionRequest.memo,
-                        msg: [...transactionRequest.msgs],
-                        signatures: [],
-                      },
-                      sequence: transactionRequest.sequence,
-                      account_number: transactionRequest.account_number,
-                      chain_id: transactionRequest.chain_id,
-                    }
-                    const signedTx = await adapter.signTransaction({
-                      txToSign: txToSign as ThorchainSignTx, // TODO: fix cosmos sdk types in hdwallet-core as they misalign and require casting,
-                      wallet,
-                    })
-                    const output = await adapter.broadcastTransaction({
-                      senderAddress: from,
-                      receiverAddress: tradeQuote.receiveAddress,
-                      hex: signedTx,
-                    })
-
-                    return output
-                  },
-                })
-                cancelPollingRef.current = output?.cancelPolling
-                return
-              }
-              case CHAIN_NAMESPACE.Solana: {
-                const adapter = assertGetSolanaChainAdapter(stepSellAssetChainId)
-                const from = await adapter.getAddress({ accountNumber, wallet })
-                const output = await execution.execSolanaTransaction({
-                  swapperName,
-                  tradeQuote,
-                  stepIndex: hopIndex,
-                  slippageTolerancePercentageDecimal,
-                  from,
-                  signAndBroadcastTransaction: async (txToSign: SolanaSignTx) => {
-                    const signedTx = await adapter.signTransaction({
-                      txToSign,
-                      wallet,
-                    })
-                    const output = await adapter.broadcastTransaction({
-                      senderAddress: from,
-                      receiverAddress: tradeQuote.receiveAddress,
-                      hex: signedTx,
-                    })
-
-                    return output
-                  },
-                })
-                cancelPollingRef.current = output?.cancelPolling
-                return
-              }
-              default:
-                assertUnreachable(stepSellAssetChainNamespace)
-            }
-          }
-        : skipToken,
-  })
 
   const executeTrade = useCallback(() => {
     if (!wallet) throw Error('missing wallet')
@@ -644,5 +469,5 @@ export const useTradeExecution = (
     isThorFreeTrade,
   ])
 
-  return { executeTrade, quoteNetworkFeesCryptoBaseUnitQuery }
+  return executeTrade
 }
